requirements.txt
python-snap7
requests
fastapi[all]